refactor into utility functions
