+++ packages/fiber/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@react-three/fiber",
-  "version": "8.0.0-alpha-08",
+  "version": "8.0.0-alpha-09",
   "description": "A React renderer for Threejs",
   "keywords": [
     "react",
