+++ packages/fiber/src/core/events.ts
@@ -1,7 +1,7 @@
 import * as THREE from 'three'
-import type { UseStore } from 'zustand'
-import type { Instance } from './renderer'
-import type { RootState } from './store'
+import { UseStore } from 'zustand'
+import { Instance } from './renderer'
+import { RootState } from './store'
 
 export interface Intersection extends THREE.Intersection {
   eventObject: THREE.Object3D
