+++ packages/fiber/src/core/hooks.ts
@@ -1,7 +1,7 @@
 import * as THREE from 'three'
 import * as React from 'react'
 import { StateSelector, EqualityChecker } from 'zustand'
-import type { GLTF } from 'three/examples/jsm/loaders/GLTFLoader'
+import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader'
 import { useAsset } from 'use-asset'
 import { context, RootState, RenderCallback } from './store'
 import { stringLiteral } from '@babel/types'
